rename keys to codes for consistency with code on master
